requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.8.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
reportlab>=4.0.0
pandas>=2.0.0
pathlib2>=2.3.7